requirements.txt
# Core Framework
Flask==3.0.0
Jinja2==3.1.2

# Database
SQLAlchemy==2.0.23
alembic==1.13.1
Flask-SQLAlchemy==3.1.1

# Authentication & Security
Flask-Login==0.6.3
argon2-cffi==23.1.0
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Image Processing
Pillow==10.4.0

# Environment & Config
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
ruff==0.1.7

# Production WSGI
gunicorn==21.2.0
